re-arranging FFI APIs subset into dedicated modules. re-factoring unit tests to reflect these changes
