+++ platforms/hermes-python/hermes_python/api/ffi/dialogue.py
@@ -1,63 +1,13 @@
-from ctypes import POINTER, c_void_p, c_char_p, byref
-from ..ffi.ontology import CProtocolHandler, CMqttOptions
-from ..ffi.ontology.facades import CDialogueFacade, CSoundFeedbackFacade
-from ..ffi.ontology.dialogue import CIntentMessage, CSessionStartedMessage, \
+from ctypes import POINTER, c_char_p, byref, c_void_p
+
+from ...ffi import utils
+from ...ffi.wrappers import ffi_function_callback_wrapper
+from ...ffi.ontology.facades import CDialogueFacade
+from ...ffi.utils import hermes_protocol_handler_dialogue_facade, hermes_drop_dialogue_facade
+from ...ffi.ontology.dialogue import CIntentMessage, CSessionStartedMessage, \
     CSessionQueuedMessage, CSessionEndedMessage, CIntentNotRecognizedMessage
-from ..ffi.utils import hermes_protocol_handler_new_mqtt_with_options, \
-    hermes_protocol_handler_dialogue_facade, hermes_drop_dialogue_facade, hermes_protocol_handler_sound_feedback_facade, \
-    hermes_drop_sound_feedback_facade
-from ..ffi.wrappers import ffi_function_callback_wrapper
-from ..ffi import utils, lib
-from ..ontology.dialogue import IntentMessage, SessionStartedMessage, SessionQueuedMessage, SessionEndedMessage, \
+from ...ontology.dialogue import IntentMessage, SessionStartedMessage, SessionQueuedMessage, SessionEndedMessage, \
     IntentNotRecognizedMessage
-
-
-class FFI(object):
-    def __init__(self, use_json_api=True, rust_logs_enabled=False):
-        self.use_json_api = use_json_api
-        self.rust_logs_enabled = rust_logs_enabled
-
-        # API Subsets
-        self.dialogue = DialogueFFI(use_json_api)
-        self.sound_feedback = SoundFeedBackFFI(use_json_api)
-        self.injection = InjectionFFI(use_json_api)
-
-        self._protocol_handler = POINTER(CProtocolHandler)()
-
-    def establish_connection(self, mqtt_options):
-        c_mqtt_options = CMqttOptions.from_repr(mqtt_options)
-
-        hermes_protocol_handler_new_mqtt_with_options(byref(self._protocol_handler), byref(c_mqtt_options))
-        self.initialize_facades()
-
-        if self.rust_logs_enabled:
-            lib.hermes_enable_debug_logs()
-
-    def initialize_facades(self):
-        self.dialogue.initialize_facade(self._protocol_handler)
-        self.sound_feedback.initialize_facade(self._protocol_handler)
-
-    def release_facades(self):
-        self.dialogue.release_facade()
-        self.sound_feedback.release_facade()
-
-    def release_connection(self):
-        self._protocol_handler = POINTER(CProtocolHandler)()
-        self.release_facades()
-
-    def _call_foreign_function(self, foreign_function_name, function_argument):
-        if self.use_json_api:
-            foreign_function_name = foreign_function_name + "_json"
-            a_json_string = str(function_argument)  # function_argument should be a dict.
-            ptr_to_foreign_function_argument = c_char_p(a_json_string.encode('utf-8'))
-        else:
-            function_argument = function_argument.into_c_repr()
-            ptr_to_foreign_function_argument = byref(function_argument)
-
-        getattr(utils, foreign_function_name)(
-            self._facade,
-            ptr_to_foreign_function_argument
-        )
 
 
 class DialogueFFI(object):
@@ -241,52 +191,3 @@ class DialogueFFI(object):
             self._facade,
             ptr_to_foreign_function_argument
         )
-
-
-# API subset stubs
-
-
-class InjectionFFI(object):
-    def __init__(self, use_json_api=True):
-        pass
-
-
-class SoundFeedBackFFI(object):
-    def __init__(self, use_json_api=True):
-        self.use_json_api = use_json_api
-        self._facade = POINTER(CSoundFeedbackFacade)()
-
-    def initialize_facade(self, protocol_handler):
-        hermes_protocol_handler_sound_feedback_facade(protocol_handler, byref(self._facade))
-
-    def release_facade(self):
-        hermes_drop_sound_feedback_facade(self._facade)
-        self._facade = POINTER(CSoundFeedbackFacade)()
-
-    def _call_foreign_function(self, foreign_function_name, function_argument):
-        if self.use_json_api:
-            foreign_function_name = foreign_function_name + "_json"
-            a_json_string = str(function_argument)  # function_argument should be a dict.
-            ptr_to_foreign_function_argument = c_char_p(a_json_string.encode('utf-8'))
-        else:
-            function_argument = function_argument.into_c_repr()
-            ptr_to_foreign_function_argument = byref(function_argument)
-
-        getattr(utils, foreign_function_name)(
-            self._facade,
-            ptr_to_foreign_function_argument
-        )
-
-    def publish_toggle_on(self, message):
-        self._call_foreign_function(
-            'hermes_sound_feedback_publish_toggle_on',
-            message
-        )
-        return self
-
-    def publish_toggle_off(self, message):
-        self._call_foreign_function(
-            'hermes_sound_feedback_publish_toggle_off',
-            message
-        )
-        return self
