+++ platforms/hermes-python/hermes_python/api/ffi/injection.py
@@ -1,0 +1,3 @@
+class InjectionFFI(object):
+    def __init__(self, use_json_api=True):
+        pass
